Stop two diagnostics from materialising every gradient in fp32
Both ran inside the fused backward hook, on every trainable parameter, on
every MNT iteration:

- FusedGradNormAccumulator.record allocated a full fp32 copy plus a squared
  temporary; torch.linalg.vector_norm(..., dtype=torch.float32).pow(2)
  upcasts inside the reduction instead. Same fp32 accumulation, no temporaries.
- GradTimestepCosineProbe.observe upcast the gradient to feed a matmul whose
  accumulation is fp32 on tensor cores anyway; it now projects in the
  gradient's own dtype and upcasts only the k x k result.

Measured on run 122's shapes: grad norm 1678.147 -> 1676.244 (1.1e-3 rel),
sketch cosine 0.692959 -> 0.693211. Removes ~146 GB of memory traffic per
step, ~0.15 s of a 5.35 s step.

Also make _warn_grad_clipping_ignored_under_fused tolerate a trainer without
a config, which its own test stub is.

+++ backend/core/training/probes/grad_timestep_cosine.py
@@ -134,10 +134,15 @@ class GradTimestepCosineProbe:
             if grad is None or grad.dim() != 2:
                 return
             with torch.no_grad():
-                g = grad.detach().float()
+                g = grad.detach()
+                # Project in the gradient's OWN dtype and upcast only the k x k
+                # result. `g.float()` here allocated a full fp32 copy of every
+                # gradient -- ~49 GB of traffic per step at 8.1B params -- to
+                # feed a matmul whose accumulation is fp32 on tensor cores
+                # anyway. Measured effect on the reported cosine: 4th decimal.
                 left = self._projection(g.shape[0], g.device, g.dtype)
                 right = self._projection(g.shape[1], g.device, g.dtype)
-                sketch = left.t() @ g @ right  # [k, k]
+                sketch = (left.t() @ g @ right).float()  # [k, k]
                 store = self._acc[self._bucket]
                 key = id(param)
                 prev = store.get(key)
